+++ sdk/identity/identity/src/index.ts
@@ -8,7 +8,6 @@ export { AuthenticationRecord } from "./msal/types";
 export { AuthenticationRequired } from "./msal/errors";
 export { serializeAuthenticationRecord, deserializeAuthenticationRecord } from "./msal/utils";
 export { TokenCredentialOptions } from "./client/identityClient";
-export { PersistentCredentialOptions } from "./credentials/persistentCredentialOptions";
 export { InteractiveCredentialOptions } from "./credentials/interactiveCredentialOptions";
 export { TokenCachePersistenceOptions } from "./tokenCache/persistencePlatforms";