+++ sdk/synapse/arm-synapse/src/models/keysMappers.ts
@@ -46,6 +46,7 @@ export {
   KeyInfoListResult,
   LibraryInfo,
   LibraryRequirements,
+  LibraryResource,
   LinkedIntegrationRuntimeKeyAuthorization,
   LinkedIntegrationRuntimeRbacAuthorization,
   LinkedIntegrationRuntimeType,